mpi3mr: Support for NVMe DIF-DIX defined in MPI Rev S
- Module parameters enable_dif and enable_dix are added and
by default enable_dif is set to true and enable_dix is
set to false. When enable_dix is set to true,
enable_dif also gets set to true.

- Support for IP_CHECKSUM as guard check type is removed.

- Support for DIX0 is removed.

- When the host metadata bit is set and when the controller
is AV1 B0 or above, the driver is modified to check each
READ/WRITE I/O issued to the NVMe DIF enabled drives formatted
with separate metadata and handle the I/Os as below:

- If the command is READ/WRITE and if xPROTECT is set to 0,
irrespective of whether HostMetaData is provided or not the
driver will send the SCSI IO request with DMAOperation bit
in Flags, MetaSGLValid bit in MsgFlags cleared to 0 and
DivertToFirmware bit in MsgFlags set to 1.

- If the command is READ/WRITE and if xPROTECT is set to 1,
and if HostMetaData is not provided then the driver will
send the SCSI IO request with DMAOperation bit in Flags,
MetaSGLValid bit in MsgFlags cleared to 0 and
DivertToFirmware bit in MsgFlags set to 1.

- If the command is READ/WRITE and if xPROTECT is set to 1,
and if HostMetaData is provided then the driver will send
the SCSI IO request with DMAOperation bit in Flags,
MetaSGLValid bit in MsgFlags set to 1 with MetaSGE and
DivertToFirmware bit in MsgFlags cleared to 0.

- For non DIF NVMe drives and when only DIF is enabled in
the driver, the driver sends the I/O “as is” irrespective
of xPROTECT values and passes the I/O to the controller
with the DMAOperation bit in Flags, MetaSGLValid bit and
DivertToFirmware bit in MsgFlags in the SCSI IO request
cleared to 0.