mpi3mr: Add support for PCIe Managed Switch SES device
The SAS4 Controller firmware exposes the SES devices
in Managed PCIe Switch as a PCIe Device Type SCSI Device
(MPI3_DEVICE0_PCIE_DEVICE_INFO_TYPE_SCSI_DEVICE).
Driver is enhanced to handle this device type and
expose the device to the upper layers.

Also, included below generic changes,
- Increase internal command timeouts to 60 seconds,
- Enabled 16 device removal handshake processing
 in parallel in the device removal handshake infrastructure.
- NVMe Encapsulated Commands issued through the IOCTL interface
with PRP format will be returned to the application with the
appropriate error code when the device to which the command issued
has zero/invalid page size.